chore: model verifier internals for knip and regenerate docs map
