+++ docs/docs_map.md
@@ -8638,6 +8638,18 @@ Do not edit it by hand; run `pnpm docs:map:gen`.
   - H2: License
   - H2: Related
 
+## reference/database-schemas.md
+
+- Route: /reference/database-schemas
+- Headings:
+  - H2: Database layout
+  - H2: Versioning contract
+  - H2: Agent schema history
+  - H2: State schema history
+  - H2: Integrity checks
+  - H2: Downgrades are unsupported
+  - H3: Example: agent schema 11 to 9
+
 ## reference/device-models.md
 
 - Route: /reference/device-models
